+++ packages/loader/src/ktx2/transcoder/KhronosWorkerCode.ts
